Refactor arp_spec_helper to two files in spec/support folder
