+++ spec/controllers/admin/accounts_controller_spec.rb
@@ -1,5 +1,4 @@
-require File.expand_path(File.dirname(__FILE__) + '/../../rails_helper')
-require File.expand_path(File.dirname(__FILE__) + '/../../arp_spec_helper')
+require 'rails_helper'
 
 describe Admin::AccountsController do
   before(:context) do
