+++ spec/controllers/protected_controller_spec.rb
@@ -1,4 +1,4 @@
-require File.dirname(__FILE__) + '/../rails_helper'
+require 'rails_helper'
 
 describe ProtectedController do
 end
