+++ spec/lib/arp/migrate_to_devise_password_spec.rb
@@ -1,5 +1,4 @@
 require 'rails_helper'
-require 'arp_spec_helper'
 
 RSpec.describe MigrateToDevisePassword do
