+++ spec/mailers/vm_spec.rb
@@ -1,5 +1,4 @@
-require File.dirname(__FILE__) + '/../rails_helper'
-require File.dirname(__FILE__) + '/../arp_spec_helper'
+require 'rails_helper'
 
 RSpec.describe Mailers::Vm, type: :mailer do
   context 'with account and VM' do
