+++ spec/models/bandwidth_quota_spec.rb
@@ -1,4 +1,4 @@
-require File.dirname(__FILE__) + '/../rails_helper'
+require 'rails_helper'
 
 context 'BandwidthQuota class' do
   before do
