+++ spec/models/login_spec.rb
@@ -1,5 +1,4 @@
-require File.dirname(__FILE__) + '/../rails_helper'
-require File.dirname(__FILE__) + '/../arp_spec_helper'
+require 'rails_helper'
 
 context Login do
   context 'with a username and password pair' do
