+++ spec/models/service_code_spec.rb
@@ -1,4 +1,4 @@
-require File.dirname(__FILE__) + '/../rails_helper'
+require 'rails_helper'
 
 describe ServiceCode do
   specify 'should count two ServiceCodes' do
