+++ spec/models/stripe_event_spec.rb
@@ -1,5 +1,4 @@
 require 'rails_helper'
-require 'arp_spec_helper'
 
 RSpec.describe StripeEvent, type: :model do
   context 'with event' do
