+++ spec/system/admin/create_account_spec.rb
@@ -1,5 +1,4 @@
 require 'rails_helper'
-require 'arp_spec_helper'
 
 RSpec.describe 'Admin - Create an account' do
