+++ spec/system/forgot_password_spec.rb
@@ -1,5 +1,4 @@
 require 'rails_helper'
-require 'arp_spec_helper'
 
 RSpec.describe 'Forgot Password' do
